Use own tmux server for catmux (#15)
* Use own tmux server for catmux

Tmux uses the environment where the server is started in. If we already
have a tmux server running when the catmux session is created, that
environment is re-used. The changes from this commit allow setting up
an environment and then start catmux with a new tmux-server that will use
that environment (as long as there is not already a running server with the
given name).